Deletes all code related to loading and building
This vastly simplifies the implementation and puts
the responsibility on the loading to deal with
this. It means that graph statistics can
(for example) be calculated in Hadoop/MR before
hand or some other means. Since this is a once per
graph thing, a bit of time overhead is perfectly
acceptable.